Made screen composable free of viewmodel Moved creation of viewmodel inside navigation destination composable so that screen composable is simple and only accepts UI state and lambda events parameter.
