+++ app/src/main/java/com/whitecatlabs/grocery/main/ui/main/MainPage.kt
@@ -26,8 +26,6 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.whitecatlabs.grocery.R
 import com.whitecatlabs.grocery.main.databse.dao.CategoryWithSelected
 import com.whitecatlabs.grocery.main.ui.AppBar
@@ -35,15 +33,6 @@ import com.whitecatlabs.grocery.main.ui.theme.AppTheme
 
 @Composable
 fun MainScreen(
-    viewModel: MainViewModel = hiltViewModel(),
-    onEvent: (MainContract.Event) -> Unit
-) {
-    val viewState = viewModel.uiState.collectAsStateWithLifecycle().value
-    MainPage(viewState) { onEvent(it) }
-}
-
-@Composable
-private fun MainPage(
     viewState: MainContract.ViewState,
     modifier: Modifier = Modifier,
     onEvent: (MainContract.Event) -> Unit
@@ -176,6 +165,6 @@ fun Grocery(
 @Composable
 private fun MainPagePreview() {
     AppTheme {
-        MainPage(viewState = MainContract.ViewState.Error) {}
+        MainScreen(viewState = MainContract.ViewState.Error) {}
     }
 }
